pyproject.toml
[project]
name = "docling-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.28",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
]
